move generic setup into base_test file
