+++ arches_project/tests/test_ui/test_settings.py
@@ -11,30 +11,14 @@ from arches_project.ui.edit_resource_replace_confirmation_dialog import EditReso
 
 from arches_project.arches_project import ArchesProject
 from arches_project.core.views.stylesheets import PluginStylesheets
+from arches_project.tests.base_test import ArchesQGISTestCase
+
 from utils.utilities import get_qgis_app
 CANVAS, PARENT, IFACE, QGIS_APP = get_qgis_app()
 
 
-class SettingsTabTests(unittest.TestCase):
+class SettingsTabTests(ArchesQGISTestCase):
     """Simple test to check dialog works."""
-
-    def setUp(self):
-        """Runs before each test."""
-        self.dlg = ArchesProjectDialog()            
-        self.dlg_resource_creation = CreateResourceConfirmation()
-        self.dlg_edit_resource_add = EditResourceAddConfirmation()
-        self.dlg_edit_resource_replace = EditResourceReplaceConfirmation()
-        self.arches_project = ArchesProject(IFACE)
-        PluginStylesheets(self.dlg, 
-                          self.dlg_resource_creation, 
-                          self.dlg_edit_resource_add, 
-                          self.dlg_edit_resource_replace, 
-                          self.arches_project.plugin_dir, 
-                          True)
-
-    def tearDown(self):
-        """Runs after each test."""
-        self.dlg = None
 
     def test_arches_theme_button(self):
         """
